pyproject.toml
[project]
name = "reference-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=4.0.0",
    "PyPDF2>=3.0.0",
    "pandas>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0"
]
